EthPricePredictor/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eth-price-predictor"
version = "0.1.0"
description = "Clean, scriptable pipeline for Ethereum price forecasting"
readme = "README.md"
authors = [{name = "Danilo Kacanski"}]
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.23",
    "pandas>=1.5",
    "scikit-learn>=1.3",
    "statsmodels>=0.14",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
lstm = ["tensorflow>=2.12"]

[project.scripts]
eth-price-pipeline = "scripts.run_pipeline:main"
